GDB Server: Remove 'fake_thread' workaround for OS threads. [DO NOT UPSTREAM]
The 'fake_thread' workaround resolves an issue with respect to
GDB debugging multiple OS threads which is outside the scope
of OpenOCD. The workaround breaks HW CPU thread support.
Since OpenOCD only supports debugging from a CPU core
perspective and not OS Threads, it is unnecessary to include
this workaround. My recommendation is to remove the workaround
since it isn't needed and breaks HW CPU thread functionality.

Tested on an Ampere eMAG8180 and Quicksilver silicon.

Change-Id: I7b09d56727a860fb86652bb3ff1e9c2f8cf6fb7b
Signed-off-by: Daniel Goehring <dgoehrin@os.amperecomputing.com>