target/aarch64: preserve memaccess tck setting
The default non-zero memaccess tck cycle setting
is preserved during the first examination. The
dap memaccess command can be used to change the
default if a lower cycle count is desired. This
change provides use of a larger, conservative
initial tck cycle count which is useful in
support of certain emulation environments.

Testing performed on ARM ARES 4-core platform.

Change-Id: I89501f98186e1dcd5f6afa7da8a202c6474588b6
Signed-off-by: Kevin Burke <kevinb@os.amperecomputing.com>
Signed-off-by: Daniel Goehring <dgoehrin@os.amperecomputing.com>